app/requirements.txt
huggingface_hub
matplotlib
numpy
opencv_python
opencv_python_headless
pandas
Pillow
protobuf
segment_anything
segmenteverygrain
streamlit
tensorflow
torch
torchvision
tqdm
streamlit 



